Test theories for motor current draw
They are commented out, and in the hardware, the motors were set to public instead of private. Without that, they are not able to be called in a teleop. So, that should not change anything and it will allow easy calling of the current draw on the motors. I also added the get pinpoint position because if the turret or robot position is drifting, it could be a bad pinpoint V2, or bad odo pods, which would need to be tested further by plugging into a control hub and getting the reading from there. And, in the getting currents, I added addLines to make it easier to read on the driver station.